refactor: merge differential loader module into core differential loader
  * move DifferentialMapLoaderModule declaration/implementation into differential_map_loader hpp/cpp
  * remove obsolete differential_map_loader_module hpp/cpp files
  * update node and test includes to use differential_map_loader.hpp
  * preserve existing behavior and ROS service wiring
