+++ map/autoware_map_loader/CMakeLists.txt
@@ -12,7 +12,6 @@ ament_auto_add_library(pointcloud_map_loader_node SHARED
   src/pointcloud_map_loader/partial_map_loader.cpp
   src/pointcloud_map_loader/partial_map_loader_module.cpp
   src/pointcloud_map_loader/differential_map_loader.cpp
-  src/pointcloud_map_loader/differential_map_loader_module.cpp
   src/pointcloud_map_loader/selected_map_loader.cpp
   src/pointcloud_map_loader/selected_map_loader_module.cpp
   src/pointcloud_map_loader/utils.cpp
