+++ map/autoware_map_loader/src/pointcloud_map_loader/differential_map_loader.hpp
@@ -17,9 +17,13 @@
 
 #include "utils.hpp"
 
+#include <rclcpp/rclcpp.hpp>
+
 #include <autoware_map_msgs/msg/area_info.hpp>
+#include <autoware_map_msgs/srv/get_differential_point_cloud_map.hpp>
 
 #include <map>
+#include <memory>
 #include <string>
 #include <vector>
 
@@ -42,6 +46,31 @@ struct DifferentialMapLoadPlan
 DifferentialMapLoadPlan create_differential_map_load_plan(
   const autoware_map_msgs::msg::AreaInfo & area_info, const std::vector<std::string> & cached_ids,
   const std::map<std::string, PCDFileMetadata> & pcd_file_metadata_dict);
+
+class DifferentialMapLoaderModule
+{
+  using GetDifferentialPointCloudMap = autoware_map_msgs::srv::GetDifferentialPointCloudMap;
+
+public:
+  explicit DifferentialMapLoaderModule(
+    std::map<std::string, PCDFileMetadata> pcd_file_metadata_dict);
+
+  // Backward-compatible constructor used by existing tests.
+  explicit DifferentialMapLoaderModule(
+    rclcpp::Node * node, std::map<std::string, PCDFileMetadata> pcd_file_metadata_dict);
+
+  [[nodiscard]] bool create_response(
+    GetDifferentialPointCloudMap::Request::SharedPtr req,
+    GetDifferentialPointCloudMap::Response::SharedPtr res) const;
+
+private:
+  std::map<std::string, PCDFileMetadata> all_pcd_file_metadata_dict_;
+  rclcpp::Service<GetDifferentialPointCloudMap>::SharedPtr get_differential_pcd_maps_service_;
+
+  [[nodiscard]] bool on_service_get_differential_point_cloud_map(
+    GetDifferentialPointCloudMap::Request::SharedPtr req,
+    GetDifferentialPointCloudMap::Response::SharedPtr res) const;
+};
 }  // namespace autoware::map_loader
 
 #endif  // POINTCLOUD_MAP_LOADER__DIFFERENTIAL_MAP_LOADER_HPP_
