+++ map/autoware_map_loader/src/pointcloud_map_loader/pointcloud_map_loader_node.hpp
@@ -15,7 +15,7 @@
 #ifndef POINTCLOUD_MAP_LOADER__POINTCLOUD_MAP_LOADER_NODE_HPP_
 #define POINTCLOUD_MAP_LOADER__POINTCLOUD_MAP_LOADER_NODE_HPP_
 
-#include "differential_map_loader_module.hpp"
+#include "differential_map_loader.hpp"
 #include "partial_map_loader_module.hpp"
 #include "pointcloud_map_loader.hpp"
 #include "selected_map_loader_module.hpp"
