+++ map/autoware_map_loader/test/test_differential_map_loader_module.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../src/pointcloud_map_loader/differential_map_loader_module.hpp"
+#include "../src/pointcloud_map_loader/differential_map_loader.hpp"
 
 #include <rclcpp/rclcpp.hpp>
